cast cast cast

+++ server/src/net/handlers/portal/capture.rs
@@ -108,7 +108,7 @@ pub fn handle(conn: &mut Connection, msg: Capture) {
             let now = UTC::now();
             db_conn.execute("UPDATE portals SET captured = $1, captured_by = $2, umonster_on = $3,
             monster_on = $7, duration = $4, catching_end = $5 WHERE id = $6",
-            &[&now, &user.id, &umonster.id, &duration, &now.checked_add(Duration::seconds(duration as i64)).unwrap(), &portal.id, &monster.id]).unwrap();
+            &[&now, &user.id as i32, &umonster.id as i32, &duration as i32, &now.checked_add(Duration::seconds(duration as i64)).unwrap(), &portal.id as i32, &monster.id as i32]).unwrap();
             rmp_map![
                 "state" => "vacant",
             ]
